lib: reduce memory allocation when processing large config transactions
Remove the xpath field from the nb_config_cb structure in order
to reduce its size. This allows the northbound to spend less time
allocating memory during the processing of large configuration
transactions.

To make this work, use yang_dnode_get_path() to obtain the xpath
from the dnode field whenever necessary. This makes the code a
little bit more verbose but it's a good trade-off in this case.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>